Add support for remote crontab command
I've added support for running crontab on the remote machine since this
is the first remote, post-deploy command I've needed.

I opted to provide specific commands in the hope that others would be
less likely to try to try to use this as an attack vector. However it
clearly still is a mechanism by which anyone with access to your repo
config could use to arbitrarily execute commands.

For those who would like to refactor this to provide some proper
escaping/validation/etc, please do so. I may get to it later, but my
environment is generally trusty.